The Dummy Data class has been created that refers to different tuples of each of the existing entities to insert them into the database (Preloaded data already in the database).
The SnapMapDatabaseInitializer class has been created that adds the tuples created in DummyData to each of the database tables.

The insertAll method has been added to each of the entities to be able to insert the entire list of tuples in SnapMapDatabaseInitializer.

All the necessary files have been created to be able to view the comments of a photo (xml files added, code implemented in CommentsFragment, the CommentsAdapter class has been created...)